# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Swagger MMS",
    "description": "locpro's magic",
    "termsOfService": "",
    "contact": {
      "email": "nhoxloctran125@gmail.com"
    },
    "license": {
      "name": "No license",
      "url": ""
    },
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "http://localhost:5000/",
      "description": "localhost"
    },
    {
      "url": "https://capstone-api-mms.herokuapp.com/",
      "description": "Heroku app"
    }
  ],
...
  "paths": {
...
    "/api/sendo/category/": {
      "get": {
        "tags": [
          "sendoAPI"
        ],
        "summary": "Search category by lazada categoryName",
        "operationId": "searchCategory",
        "security": [
          {
            "bearerAuth": [],
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "required": true,
            "description": "string of category to search",
            "schema": {
              "type": "string",
              "example": "Áo khoác Denim nam"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful"
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Pet not found"
          },
          "405": {
            "description": "Validation exception"
          }
        }
      }
    },
...
    "/categories": {
      "get": {
        "tags": [
          "categories"
        ],
        "summary": "get list categories in database",
        "operationId": "getListCategories",
        "security": [
          {
            "bearerAuth": [],
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "idpath[]",
            "required": false,
            "schema": {
              "type": "array",
              "items": []
            },
            "description": "categoryId must be primary category"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful"
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "405": {
            "description": "Validation exception"
          }
        }
      },
      "post": {
        "tags": [
          "categories"
        ],
        "summary": "create category in database",
        "operationId": "createCategory",
        "security": [
          {
            "bearerAuth": [],
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {},
              "example": {
                "idpath": [],
                "namepath": [],
                "category_id": 99999,
                "var": false,
                "name": "locpro",
                "leaf": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful"
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "405": {
            "description": "Validation exception"
          }
        }
      }
    },
    "/categories/{categoryId}": {
      "get": {
        "tags": [
          "categories"
        ],
        "summary": "get category by id in database",
        "operationId": "getCategoryById",
        "security": [
          {
            "bearerAuth": [],
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "categoryId",
            "required": true,
            "schema": {
              "type": "number",
              "example": 10100180
            },
            "description": "categoryId must be primary category"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful"
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "405": {
            "description": "Validation exception"
          }
        }
      },
      "patch": {
        "tags": [
          "categories"
        ],
        "summary": "update category in database",
        "operationId": "updateCategory",
        "security": [
          {
            "bearerAuth": [],
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "categoryId",
            "required": true,
            "schema": {
              "type": "number",
              "example": "99999"
            },
            "description": ""
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {},
              "example": {
                "idpath": [],
                "namepath": [],
                "var": false,
                "name": "locpro1",
                "leaf": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful"
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "405": {
            "description": "Validation exception"
          }
        }
      },
      "delete": {
        "tags": [
          "categories"
        ],
        "summary": "delete category by id in database",
        "operationId": "deleteCategory",
        "security": [
          {
            "bearerAuth": [],
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "categoryId",
            "required": true,
            "schema": {
              "type": "number",
              "example": "99999"
            },
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "Successful"
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "405": {
            "description": "Validation exception"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "in": "header",
        "name": "Authorization",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Platform-Token"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}